LayoutTests/http/tests/security/contentSecurityPolicy/script-src-self-blocked-02-expected.txt

CONSOLE MESSAGE: Refused to load script from 'http://127.0.0.1:8080/security/contentSecurityPolicy/resources/script.js' because of Content-Security-Policy.



--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
PASS
